% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_area.R
\name{calc_area}
\alias{calc_area}
\title{Caculate the no-data zone areas}
\usage{
calc_area(xdat, ydat)
}
\arguments{
\item{xdat}{a vector of the independent data}

\item{ydat}{a vector of the dependent data}
}
\value{
a list of the no-data zone areas
}
\description{
This function supports the other sobir functions by calculating the no-data zone areas
}
\examples{
a = rnorm(100,0,1)
b = rnorm(100,0,1)
calc_area(a,b)
}
