% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bpts.R
\name{extract_bpts}
\alias{extract_bpts}
\title{Extract boundary points}
\usage{
extract_bpts(xdat, ydat)
}
\arguments{
\item{xdat}{a vector of the independent data}

\item{ydat}{a vector of the dependent data}
}
\value{
a data frame of the two vectors with the boundary points and other relevant data labelled.
}
\description{
extract_bpts identifies the boundary points to a scatterplot and labels the
relevant data for further analysis in the sobir package.
}
\examples{
a = rnorm(100,0,1)
b = rnorm(100,0,1)
extract_bpts(a,b)
}
