% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodyAfrica.R
\docType{data}
\name{WoodyAfrica}
\alias{WoodyAfrica}
\title{Woody cover in Africa}
\format{A data frame with 854 rows and 2 variables}
\source{
\url{https://doi.org/10.1038/nature04070}
}
\usage{
WoodyAfrica
}
\description{
Woody cover percentage in Africa relative to mean annual precipitation (mm), 
as published by Sankaran et al. 2005 in Nature.
}
\details{
MAP = Mean annual precipitation, in mm
Cover = Woody cover, in %
}
\references{
Sankaran et al. (2005) Determinants of woody cover in African savannas. Nature 438, p846–849.
}
\keyword{datasets}
