% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodyTowoomba.R
\docType{data}
\name{WoodyTowoomba}
\alias{WoodyTowoomba}
\title{Woody cover in Towoomba}
\format{}
\source{
\url{https://doi.org/10.1371/journal.pone.0179848}
}
\usage{
WoodyTowoomba
}
\description{
Tree abundance in Towoomba Agricultural Development Center, South Africa relative to 
four soil fertilisation treatments and 16 soil nutrient and element concentrations, 
as published by Mills et al. 2017 in PLoS ONE.
}
\references{
Mills et al. (2017) Effects of anabolic and catabolic nutrients on woody plant encroachment after long-term experimental fertilization in a South African savanna. PLoS ONE 12(6), p1-24.
}
\keyword{datasets}
