% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_area.R
\name{perm_area}
\alias{perm_area}
\title{Calculate the permuted area}
\usage{
perm_area(xdat, ydat, nsim, boundary = "topl", method = "auto")
}
\arguments{
\item{xdat}{a vector of the independent data}

\item{ydat}{a vector of the dependent data}

\item{nsim}{the number of simulations to run}

\item{boundary}{character string indicating the boundary to test (default is "topl"). Possible values are "topl" (top-left), "topr" (top-right), "botl" (bottom-left), "botr" (bottom-right) or "all".}

\item{method}{character string indicating computation method (default is "auto"). Possible values are "exact", "approximate" or "auto".}
}
\value{
a perm table that can be plotted directly using perm_plot()
}
\description{
perm_area calculates the no-data zone areas for each permutation of the data simulated nsim times.
}
\examples{
a = rnorm(100,0,1)
b = rnorm(100,0,1)
perm_area(a,b,10)
}
